requirements.txt
# Donut Core Dependencies
transformers>=4.11.3
timm
datasets[vision]
pytorch-lightning>=1.6.4
nltk
sentencepiece
zss
sconf>=0.2.3

# PyTorch (Install separately based on your CUDA version)
# torch>=1.11.0
# torchvision>=0.12.0

# PDF and Image Processing
pdf2image>=1.16.0
PyPDF2>=3.0.0
pypdfium2>=4.0.0
Pillow>=9.0.0
opencv-python>=4.5.0

# API Framework
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
jsonlines>=3.0.0

# Evaluation and Metrics
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
